Fixed PhantomJS result for Object#__proto__
PH's __proto__ implementation does not have a visible property
descriptor.

Also decided to add guards for the ES5 methods used in the test. Since
Object.getOwnPropertyDescriptor is the only way to guarantee that it is
implemented as a get/set property, we might as well fail the test if
it's absent. (Incidentally, everything that fully supports
Object.getOwnPropertyDescriptor also supports Object.create.)