Reverted the spread/iterator test for now.
I decided that until the table has a way to refer to later tests'
results in earlier ones, then including almost the entire code of a
later test, just to check cross-compatibility, is a bit much.

Instead, I've replaced it with a check that spread will throw when
applied to non-iterables, and that it will not throw for empty
iterables.

Also, WebKit does not support string spreading.