Fixed an ES5 strict test that crashes in Chrome
The test `(function f() { f = 123; })()` actually throws a SyntaxError
on evaluation, not a TypeError on execution, in the current Chrome. I'm
really not sure which is fully correct for either the ES5 or ES6 specs.
Nevertheless, I've changed that line to be eval()ed so that the entire
script doesn't crash.

Also, the link to the strict mode page is fixed.