ES6 test functions in multi-line strings
This updates the build script, allowing it to interpret ES6 code
expressed in function comments instead of concatenated strings (see #190
discussion) - which is then wrapped in try-catch, eval() and test()
structures automatically.

Also, the runtime script has been updated to unwrap those same
structures for the tooltip preview, making the resulting display cleaner
- lacking even the test() call wrapper.

Only some ES6 and non-standard tests have been converted to make use of
this format, for the moment. The remaining tests are subject to
pull-requests at this moment and will be addressed after those are
resolved.

NOTE: as this drastically affects the output of all HTML files, they
have not been included in this commit, to allow merging to pass
smoothly.