Cleaned up master.css a bit
This also changes some of the colours for the headers in each page, to
make each of them visually distinct (previously, only the non-standard
page had different header colours).

Also, it changes the "No" colour on the non-standard page from red to
blue. The reasoning is that since the features conform to no ES
standard, the judgmental, hostile red of the other pages is
inappropriate here, and a more neutral, yet distinct, shade is required.

Also removed the user-agent tooltip <script> from the non-standard
table, as it was interfering with the DOM structure expected by some of
the CSS, and didn't really contribute much.