Flagged support is now shown in header bar-graphs.
The percentage of flagged support is now shown in the browser header bars as a more transparent bar behind the normal bar.

Also, supertests containing flagged results are now marked with the fold triangle as well.

Also, merged Node and Node Harmony into a single column, treating --harmony as a flag.