Added a 'strict' result value alongside 'flag'.
This refers exclusively to features which only pass in strict mode, after failing in non-strict mode. These are treated the same as 'flag', but have their own "Strict" text and tooltip.

The strict/non-strict split tests for let and const will remain, as they are needed to show historical results which required both strict mode and the "experimental" flag.

Closes #498.