Marked various old FF and CH versions as "very obsolete".
In a potentially misguided attempt to improve performance, all but the last 10 versions of Firefox and Chrome (before stable) have been marked as "very obsolete", which means that they won't be included in the page's increasingly hefty HTML at all. Their data will of course remain in data-es6 for archival purposes, and could return if a special "historical" version of the pages was implemented in the future.