Unify the version naming convention
So far some browsers were just omitting a dot in the version string and some
replaced it by an underscore. This commit changes all those versions to use
the underscore translation. This will reduce confusion; with some environments
using rapidly increasing major version numbers it becomes hard to distinguish
"41" from "4.1" when both are translated to "41" in the browser name.

The only exception is Duktape which needs to have names maching keys matching
what Duktape.version reports.