remove all uses of document.write
This commit updates the scripts that used 'document.write' to
instead use DOM creation methods.

The call that wrote a single CSS rules was removed with the rule
being added to 'master.css'.

The 'document.write' calls were always in the currently executing
script, so the 'document.currentScript' or the last script in
'document.scripts' (or its polyfill, which was added) was used to
obtain the proper element to target with the new nodes.

The original structure of the HTML was kept.

Resolves #1289