undo == to === for three tests
This is a partial revert of 5a9fa24

While overall beneficial, it seems that using the identity operator actually breaks those three tests. We have looked into the spec and seems that all three should not pass with the identity operator in place. There might be a better way to fix this, but in the current state, those tests fail on several engines (tested on our own GraalVM 20.0, but also on FF 45, Chrome 80, Node 12.x). I thus suggest to just undo the changes for the moment.